requirements.txt
django==1.8
pycrypto
requests